Added Save option. Cleaned the code a little bit.

+++ GLoPhysX/GLoPhysX/src/glophysx/rendering/camera/editor/editor_camera.cpp
@@ -84,7 +84,7 @@ namespace GLOPHYSX {
 
 		bool EditorCamera::OnMouseScroll(MouseScrollEvent& e)
 		{
-			float delta = e.GetYOffset() * 0.1f;
+			float delta = (float)e.GetYOffset() * 0.1f;
 			MouseZoom(delta);
 			UpdateView();
 			return false;
@@ -140,7 +140,7 @@ namespace GLOPHYSX {
 			float distance = m_distance * 0.2f;
 			distance = std::max(distance, 0.0f);
 			float speed = distance * distance;
-			speed = std::min(speed, 100.0f); // max speed = 100
+			speed = std::min(speed, 100.0f);
 			return speed;
 		}
 	}
